requirements.txt
lxml==5.4.0
python-docx==1.1.2
typing_extensions==4.13.2
numpy==1.24.0
pandas==2.0.0

# Testing
pytest>=7.0.0
